move `UnifiedState`, `UnifiedControl` dtaclases to `lowered/`
