+++ openscvx/lowered/__init__.py
@@ -10,6 +10,8 @@ Classes:
     CVXPyVariables: CVXPy variables and parameters for the OCP
     ParameterDict: Dictionary that syncs parameters between JAX and CVXPy
     Dynamics: JAX-lowered dynamics functions (f, A, B)
+    UnifiedState: Unified state vector metadata for the lowered representation
+    UnifiedControl: Unified control vector metadata for the lowered representation
 """
 
 from openscvx.lowered.cvxpy_constraints import LoweredCvxpyConstraints
@@ -22,6 +24,7 @@ from openscvx.lowered.jax_constraints import (
 )
 from openscvx.lowered.parameters import ParameterDict
 from openscvx.lowered.problem import LoweredProblem
+from openscvx.lowered.unified import UnifiedControl, UnifiedState
 
 __all__ = [
     "LoweredProblem",
@@ -32,4 +35,6 @@ __all__ = [
     "CVXPyVariables",
     "ParameterDict",
     "Dynamics",
+    "UnifiedState",
+    "UnifiedControl",
 ]
